DIAC-1516 reordering pages
